requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
inquirerpy==0.3.4
jiter==0.5.0
numpy==2.3.2
outcome==1.3.0.post0
pydantic==2.9.1
pydantic_core==2.23.3
PySocks==1.7.1
python-dotenv==0.21.0
python-http-client==3.3.7
requests==2.32.3
selenium==4.24.0
sendgrid==6.11.0
sniffio==1.3.1
sortedcontainers==2.4.0
starkbank-ecdsa==2.2.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
websocket-client==1.8.0
wsproto==1.2.0

# Gmail API dependencies
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0

# Simulator dependencies
pandas==2.2.3
matplotlib==3.9.2

# Supabase for realtime database
supabase==2.22.2
